+++ setup_kitti.sh
@@ -1,8 +1,17 @@
-# Change lines below to the path where you store the dataset.
+# Change lines below.
+# Path to KITTI dataset.
 KITTI_DATA=/ais/gobi4//mren/data/kitti/object
+# Path to model storage.
+SAVE_FOLDER=/ais/gobi5/mren/results/rec-attend
+# Path to log storage.
+DASHBOARD_LOGS=/u/mren/public_html/results
+
 mkdir -p data
-ln -s $KITTI_DATA data/kitti
-# ./setup_kitti.py
+
+if [ ! \( -e "${data/kitti}" \) ]
+then
+  ln -s $KITTI_DATA data/kitti
+fi
 
 FF='results'
 if [ ! \( -e "${FF}" \) ]
@@ -18,3 +27,4 @@ then
   DASHBOARD_LOGS=/u/mren/public_html/results
   ln -s $DASHBOARD_LOGS logs
 fi
+./setup_kitti.py
